Extract rule-related functionality into dedicated src/rule module
### Module Restructuring:
- **Created `src/rule/` module** for rule processing functionality
- **Moved `agent/rule_engine.rs` → `rule/engine.rs`**
  - Core rule evaluation and decision logic
  - Pattern matching and timeout rule processing
- **Moved `agent/timeout.rs` → `rule/timeout.rs`**
  - Timeout rule monitoring functionality
- **Added `rule/mod.rs`** with proper re-exports

### Updated Import Structure:
- **Updated `main.rs`** to include new rule module
- **Fixed all import references** across codebase
  - `config/mod.rs` → `crate::rule::engine`
  - `agent/agents.rs` → `crate::rule::TimeoutMonitor`
  - `agent/monitor.rs` → `crate::rule::Monitor`
- **Removed rule modules** from `agent/mod.rs`

### Benefits:
- **Better separation of concerns** - rules vs agent management
- **Clearer domain boundaries** - rule processing is now separate
- **Improved modularity** - easier to extend rule functionality
- **All tests passing** (27 unit + 3 integration tests)
- **Clean build** with no warnings or errors

The `src/agent/` module now focuses purely on agent lifecycle management,
while `src/rule/` handles all rule evaluation and timeout logic.
